backend/requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
fonttools==4.47.0
gunicorn==20.0.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
protobuf==3.19.0
PyJWT==1.7.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.12.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.4
Werkzeug==3.0.1
zipp==3.17.0
